ML_Model/requirements.txt
absl-py==2.0.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
frozendict==2.3.8
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.1
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.18.1
# nvidia-nvjitlink-cu12==12.3.52
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
pandas-datareader==0.10.0
peewee==3.17.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
soupsieve==2.5
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.0
# triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
yfinance==0.2.31
